Some timers fixes.

+++ include/drivers/timers.h
@@ -111,7 +111,7 @@ int timer_setup(uint32_t timer_num, struct timer_config* conf);
  *   config field in timer_config struct upon timer setup)
  *   The interrupt flags are passed to the interrupt routine as argument.
  */
-void timer_on(uint32_t timer_num, uint32_t clkrate, void (*callback)(uint8_t));
+void timer_on(uint32_t timer_num, uint32_t clkrate, void (*callback)(uint32_t));
 
 /* Removes the main clock from the selected timer block */
 void timer_off(uint32_t timer_num);
